requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
ollama==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
